Added figure to show launch angle dependence
Figure shows that across the range of launch angles for a barreled ball hit at 100 mph, the error in the model (almost certainly from the stop condition) is bigger than the sensitivity of a given launch angle to gravitational difference.

Next revision will have tweaks to the baseball dynamics models to account for this.